fix(IL0323): limit fading during partial refresh
"Gates scan only inside of the partial window" instead of setting PTL_SCAN
when configuring the window for partial refreshes seems to limit degrading
of areas outside the refresh window quite a bit.
But https://github.com/CursedHardware/epd-driver-ic/blob/master/IL0323.pdf,
p. 26 states that this is not the "default".